refactor(gemini): extract temporary chat handler
